requirements.txt
python-dotenv==1.0.1
transformers==4.44.2
torch==2.4.1
elasticsearch==8.15.1
langchain==0.3.0
langchain_openai==0.2.0
langgraph==0.2.22
langchain_elasticsearch==0.3.0
streamlit==1.38.0
importlib==1.0.4